+++ spec/guard/dsl_describer_spec.rb
@@ -1,9 +1,11 @@
+# encoding: utf-8
+
 require 'spec_helper'
 
 describe Guard::DslDescriber do
 
   let(:guardfile) do
-    <<-GUARD
+    <<-GUARDFILE
       guard :test, :a => :b, :c => :d do
         watch('c')
       end
@@ -19,49 +21,58 @@ describe Guard::DslDescriber do
           watch('c')
         end
       end
-    GUARD
+    GUARDFILE
   end
 
   before do
-    @output = ''
-    Guard::UI.stub(:info) { |msg| @output << msg + "\n" }
-  end
-
-  after do
-    Guard::UI.unstub(:info)
+    stub_const 'Guard::Test', Class.new(Guard::Guard)
+    stub_const 'Guard::Another', Class.new(Guard::Guard)
   end
 
   describe '.list' do
-    it "lists the available Guards when they're declared as strings or symbols" do
-      Guard.stub(:guard_gem_names).and_return ['test', 'another', 'even', 'more']
-      described_class.list(:guardfile_contents => guardfile)
-      @output.should == <<OUTPUT
-Using inline Guardfile.
-Available guards:
-   another*
-   even
-   more
-   test*
+    let(:output) do
+      <<-OUTPUT.gsub(/^\s+|\n$/, '')
+        +------------+--------------+
+        |  Available Guard plugins  |
+        +------------+--------------+
+        | Plugin     | In Guardfile |
+        +------------+--------------+
+        | Another    | ✔            |
+        | Even       | ✘            |
+        | More       | ✘            |
+        | Test       | ✔            |
+        +------------+--------------+
+      OUTPUT
+    end
 
-See also https://github.com/guard/guard/wiki/List-of-available-Guards
-* denotes ones already in your Guardfile
-OUTPUT
+    before do
+      ::Guard.stub(:guard_gem_names).and_return %w(test another even more)
+    end
+
+    it 'lists the available Guards when they\'re declared as strings or symbols' do
+      ::Guard::DslDescriber.list(:guardfile_contents => guardfile).to_s.should eq(output)
     end
   end
 
   describe '.show' do
-    it 'shows the Guards and their options' do
-      described_class.show(:guardfile_contents => guardfile)
-      @output.should == <<OUTPUT
-Using inline Guardfile.
-(global):
-  test: a => :b, c => :d
-Group a:
-  test: x => 1, y => 2
-Group b:
-  another
+    let(:output) do
+      <<-OUTPUT.gsub(/^\s+|\n$/, '')
+       +---------+---------+--------+-------+
+       |        Guardfile structure         |
+       +---------+---------+--------+-------+
+       | Group   | Plugin  | Option | Value |
+       +---------+---------+--------+-------+
+       | Default | Test    | a      | :b    |
+       |         |         | c      | :d    |
+       | A       | Test    | x      | 1     |
+       |         |         | y      | 2     |
+       | B       | Another |        |       |
+       +---------+---------+--------+-------+
+      OUTPUT
+    end
 
-OUTPUT
+    it 'shows the Guards and their options' do
+      ::Guard::DslDescriber.show(:guardfile_contents => guardfile).to_s.should eq(output)
     end
   end